pyproject.toml
[project]
name = "fastapi-finance-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.5",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
